Improving model accuracy and update DigitClassifier output format
The step 7 of Codelab allows us to improve the model accuracy at:
https://colab.research.google.com/github/tensorflow/examples/blob/master/lite/codelabs/digit_classifier/ml/step7_improve_accuracy.ipynb

The base model was trained using images with digits in the center of the image. In this step we generate new training data adding rotation and moving digits around the image to allow the model to identify more situations.

- Adds the improved `mnist.tflite` model file to the assets folder.
- Modifies the output format of `DigitClassifier` to display the confidence as a percentage.